Fixed MegaRAM mapper so it can handle 2048K (256 pages) as well.
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@5225 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/memory/MSXMegaRam.hh
@@ -34,7 +34,7 @@ private:
 
 	std::auto_ptr<Ram> ram;
 	std::auto_ptr<Rom> rom;
-	byte numBlocks;
+	unsigned numBlocks;
 	byte maskBlocks;
 	byte bank[4];
 	bool writeMode;
